requirements.txt
asgiref==3.7.2
blinker==1.4
certifi==2023.7.22
charset-normalizer==3.2.0
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
Django==4.2.5
django-cors-headers==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.0
oauthlib==3.2.0
openapi-codec==1.3.2
pandas==2.1.1
psycopg2-binary==2.9.7
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.5
wadllib==1.3.6
zipp==1.0.0
